requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
mysqlclient>=2.2,<2.3
python-decouple>=3.8,<3.9
django-filter>=24.2,<25.0
celery>=5.4,<5.5
redis>=5.0,<5.1
djangorestframework-simplejwt>=5.3,<5.4
gunicorn>=22.0,<23.0
openai>=1.30,<2.0
pypdf>=4.0,<5.0
python-docx>=1.1,<1.2
pdfplumber>=0.11,<0.12
reportlab>=4.0,<4.1
